Move GitHubOrg (and GitHubConfig) out of @/config
